Change order_id type from u32 to u64

+++ client/test/schema.ts
@@ -320,7 +320,7 @@ export const ConsumeEventsSchema: borsh.Schema = {
 
 export const CancelOrderSchema: borsh.Schema = {
     struct : {
-        order_id: "u32",
+        order_id: "u64",
         side: "u8"
     }
 }
